many optimisations, memory and performance - alot of small gains together ftw
the tableswitch was forcing inlining of the version check, if ignored it and performance suffered.  Small gains outside of that but needs further testing.  Many 1% gets, lazy vals and other unnecessary things.  Modified TreeProxies to allow for TextJoiner to work with the builders.  Using the builders as a whole seems to be efficient enough